requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
efficientnet-pytorch=0.7.1=pypi_0
imageio=2.9.0=pypi_0
imgaug=0.4.0=pypi_0
matplotlib=3.4.3=pypi_0
munch=2.5.0=pypi_0
numpy=1.21.2=pypi_0
opencv-python=4.5.3.56=pypi_0
pandas=1.3.3=pypi_0
pillow=8.3.2=pypi_0
python=3.8.11=h12debd9_0_cpython
pytorch-lightning=1.4.9=pypi_0
scikit-learn=1.0=pypi_0
torch=1.9.1=pypi_0
torchvision=0.10.1=pypi_0
tqdm=4.62.3=pypi_0
